Modularize state & prop data processing
